requirements.txt
altgraph==0.17.3
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
black==24.4.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
convertdate==2.4.0
distlib==0.3.6
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.8.0
future==0.18.2
h11==0.14.0
hijri-converter==2.2.4
holidays==0.18
httpcore==1.0.7
httpx==0.28.1
idna==3.4
jiter==0.8.0
korean-lunar-calendar==0.3.1
macholib==1.16.3
mypy-extensions==1.0.0
openai==1.57.0
packaging==24.1
pathspec==0.12.1
pbr==5.11.0
pefile==2022.5.30
platformdirs==4.2.2
pydantic==2.10.3
pydantic_core==2.27.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyMeeus==0.5.12
python-dateutil==2.8.2
python-dotenv==1.0.1
pywin32-ctypes==0.2.0
requests==2.28.1
six==1.16.0
sniffio==1.3.1
stevedore==4.1.1
tomli==2.0.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.13
